+++ rokid_3dof/examples/rokid_test.rs
@@ -6,7 +6,6 @@ use wasm_bindgen::prelude::*;
 
 use cgmath::Rotation3;
 use font_rasterizer::{
-    camera::{Camera, CameraController},
     color_theme::ColorTheme::{self, SolarizedDark},
     context::{StateContext, WindowSize},
     instances::{GlyphInstance, GlyphInstances},
@@ -14,7 +13,10 @@ use font_rasterizer::{
     rasterizer_pipeline::Quarity,
     time::now_millis,
 };
-use ui_support::{run_support, Flags, InputResult, SimpleStateCallback, SimpleStateSupport};
+use ui_support::{
+    camera::{Camera, CameraController},
+    run_support, Flags, InputResult, SimpleStateCallback, SimpleStateSupport,
+};
 
 use log::info;
 use winit::event::{ElementState, MouseButton, WindowEvent};
