+++ sample_codes/showcase/src/lib.rs
@@ -7,17 +7,19 @@ use text_buffer::action::EditorOperation;
 use wasm_bindgen::prelude::*;
 
 use font_rasterizer::{
-    camera::{Camera, CameraAdjustment},
     color_theme::ColorTheme,
-    context::{StateContext, TextContext, WindowSize},
+    context::{StateContext, WindowSize},
     instances::GlyphInstances,
-    layout_engine::{HorizontalWorld, Model, World},
     rasterizer_pipeline::Quarity,
-    ui::{caret_char, ImeInput, TextEdit},
 };
 use ui_support::{
-    action::ActionProcessorStore, run_support, Flags, InputResult, SimpleStateCallback,
-    SimpleStateSupport,
+    action::ActionProcessorStore,
+    camera::{Camera, CameraAdjustment},
+    layout_engine::{HorizontalWorld, Model, World},
+    run_support,
+    ui::{caret_char, ImeInput, TextEdit},
+    ui_context::TextContext,
+    Flags, InputResult, SimpleStateCallback, SimpleStateSupport,
 };
 use winit::event::WindowEvent;
