+++ ui_support/examples/setup_glyphs.rs
@@ -1,13 +1,12 @@
 use font_collector::FontCollector;
 use font_rasterizer::{
-    camera::Camera,
     color_theme::ColorTheme,
     context::{StateContext, WindowSize},
     instances::GlyphInstances,
     rasterizer_pipeline::Quarity,
 };
 use instant::Instant;
-use ui_support::{run_support, Flags, InputResult, SimpleStateCallback, SimpleStateSupport};
+use ui_support::{camera::Camera, run_support, Flags, InputResult, SimpleStateCallback, SimpleStateSupport};
 use winit::event::WindowEvent;
 
 const EMOJI_FONT_DATA: &[u8] = include_bytes!("../../fonts/NotoEmoji-Regular.ttf");
